Refactor main component in smaller pieces, using styled components
